Add unstick-upgrade plugin
This is specifically to work around for bugs in Juju 1.23, 1.24.0 and
1.24.1 that may prevent state machine agents from being able to
restart in order to start an upgrade. It may also help with similar
problem in future Juju versions.

The plugin is designed to be run once the upgrade has started and the
state machines are stuck. It determines the state server addresses
using "juju api-info" and then connects to each over SSH and applies a
fix which allows the upgrade to complete.

Handling for local provider environments which need to fixed in a
slightly different way is also included.